Promote some methods up to base class reducing duplication
